+++ component/loki/source/file/file_test.go
@@ -1,3 +1,5 @@
+//go:build !race
+
 package file
 
 import (
@@ -12,6 +14,7 @@ import (
 	"github.com/grafana/agent/component"
 	"github.com/grafana/agent/component/common/loki"
 	"github.com/grafana/agent/component/discovery"
+	"github.com/grafana/agent/pkg/flow/componenttest"
 	"github.com/grafana/agent/pkg/util"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
@@ -22,35 +25,35 @@ import (
 func Test(t *testing.T) {
 	defer goleak.VerifyNone(t, goleak.IgnoreTopFunction("go.opencensus.io/stats/view.(*worker).start"))
 
-	// Create opts for component
-	opts := component.Options{
-		Logger:        util.TestFlowLogger(t),
-		Registerer:    prometheus.NewRegistry(),
-		OnStateChange: func(e component.Exports) {},
-		DataPath:      t.TempDir(),
-	}
+	ctx, cancel := context.WithCancel(componenttest.TestContext(t))
+	defer cancel()
 
-	f, err := os.CreateTemp(opts.DataPath, "example")
-	if err != nil {
-		log.Fatal(err)
-	}
+	// Create file to log to.
+	f, err := os.CreateTemp(t.TempDir(), "example")
+	require.NoError(t, err)
 	defer f.Close()
 
-	ch1, ch2 := make(chan loki.Entry), make(chan loki.Entry)
-	args := Arguments{}
-	args.Targets = []discovery.Target{{"__path__": f.Name(), "foo": "bar"}}
-	args.ForwardTo = []loki.LogsReceiver{ch1, ch2}
-
-	c, err := New(opts, args)
+	ctrl, err := componenttest.NewControllerFromID(util.TestLogger(t), "loki.source.file")
 	require.NoError(t, err)
 
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	go c.Run(ctx)
-	time.Sleep(100 * time.Millisecond)
+	ch1, ch2 := make(chan loki.Entry), make(chan loki.Entry)
+
+	go func() {
+		err := ctrl.Run(ctx, Arguments{
+			Targets: []discovery.Target{{
+				"__path__": f.Name(),
+				"foo":      "bar",
+			}},
+			ForwardTo: []loki.LogsReceiver{ch1, ch2},
+		})
+		require.NoError(t, err)
+	}()
+
+	ctrl.WaitRunning(time.Minute)
 
 	_, err = f.Write([]byte("writing some text\n"))
 	require.NoError(t, err)
+
 	wantLabelSet := model.LabelSet{
 		"filename": model.LabelValue(f.Name()),
 		"foo":      "bar",
@@ -69,6 +72,44 @@ func Test(t *testing.T) {
 		case <-time.After(5 * time.Second):
 			require.FailNow(t, "failed waiting for log line")
 		}
+	}
+}
+
+// Test that updating the component does not leak goroutines.
+func TestUpdate_NoLeak(t *testing.T) {
+	defer goleak.VerifyNone(t, goleak.IgnoreTopFunction("go.opencensus.io/stats/view.(*worker).start"))
+
+	ctx, cancel := context.WithCancel(componenttest.TestContext(t))
+	defer cancel()
+
+	// Create file to tail.
+	f, err := os.CreateTemp(t.TempDir(), "example")
+	require.NoError(t, err)
+	defer f.Close()
+
+	ctrl, err := componenttest.NewControllerFromID(util.TestLogger(t), "loki.source.file")
+	require.NoError(t, err)
+
+	args := Arguments{
+		Targets: []discovery.Target{{
+			"__path__": f.Name(),
+			"foo":      "bar",
+		}},
+		ForwardTo: []loki.LogsReceiver{},
+	}
+
+	go func() {
+		err := ctrl.Run(ctx, args)
+		require.NoError(t, err)
+	}()
+
+	ctrl.WaitRunning(time.Minute)
+
+	// Update a bunch of times to ensure that no goroutines get leaked between
+	// updates.
+	for i := 0; i < 10; i++ {
+		err := ctrl.Update(args)
+		require.NoError(t, err)
 	}
 }